move host info gathering funcs to logging utils
